Refactor architecture to separate Rust libraries
