+++ ami_cli/Cargo.toml
@@ -4,4 +4,4 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-pyo3 = { version = "0.21", features = ["auto-initialize"] }
+rust_amidatabase = { path = "../rust_amidatabase" }
