+++ rust_amidatabase/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2024"
 
 [lib]
 name = "rust_amidatabase"
-crate-type = ["cdylib"]
+path = "src/lib.rs"
 
 [dependencies]
-pyo3 = { version = "0.21", features = ["extension-module"] }
+rust_amireader = { path = "../rust_amireader" }
