+++ rust_amidatabase_py/.cargo/config.toml
@@ -5,4 +5,4 @@ offline = true
 replace-with = "vendored-sources"
 
 [source.vendored-sources]
-directory = "../rust_bitparser/vendor"
+directory = "../rust_bitparser_py/vendor"
