+++ rust_amidatabase_py/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "rust_amidatabase_py"
+version = "0.1.0"
+edition = "2024"
+
+[lib]
+name = "rust_amidatabase"
+crate-type = ["cdylib"]
+
+[dependencies]
+pyo3 = { version = "0.21", features = ["extension-module"] }
+rust_amidatabase = { path = "../rust_amidatabase" }
