+++ rust_amireader_py/.cargo/config.toml
@@ -2,4 +2,4 @@
 replace-with = "vendored-sources"
 
 [source.vendored-sources]
-directory = "../rust_bitparser/vendor"
+directory = "../rust_bitparser_py/vendor"
