+++ rust_amireader_py/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "rust_amireader_py"
+version = "0.1.0"
+edition = "2024"
+
+[lib]
+name = "rust_amireader"
+crate-type = ["cdylib"]
+
+[dependencies]
+pyo3 = { version = "0.21", features = ["extension-module"] }
+rust_amireader = { path = "../rust_amireader" }
