+++ rust_bitparser/Cargo.toml
@@ -5,7 +5,4 @@ edition = "2024"
 
 [lib]
 name = "rust_bitparser"
-crate-type = ["cdylib"]
-
-[dependencies]
-pyo3 = { version = "0.21", features = ["extension-module"] }
+path = "src/lib.rs"
